+++ Samples/basic/harfbuzzshaping/src/FontEngineInterfaceHarfBuzz.h
@@ -40,6 +40,7 @@ using Rml::FontEffectsHandle;
 using Rml::FontFaceHandle;
 using Rml::FontMetrics;
 using Rml::RenderManager;
+using Rml::Span;
 using Rml::String;
 using Rml::TextShapingContext;
 using Rml::TexturedMeshList;
@@ -55,7 +56,7 @@ public:
 	bool LoadFontFace(const String& file_name, bool fallback_face, Style::FontWeight weight) override;
 
 	/// Adds a new font face to the database using the provided family, style and weight.
-	bool LoadFontFace(const byte* data, int data_size, const String& font_family, Style::FontStyle style, Style::FontWeight weight,
+	bool LoadFontFace(Span<const byte> data, const String& font_family, Style::FontStyle style, Style::FontWeight weight,
 		bool fallback_face) override;
 
 	/// Returns a handle to a font face that can be used to position and render text. This will return the closest match